pyproject.toml
[tool.poetry]
name = "ritsu"
version = "0.1.0"
description = "Official Discord Bot for Assassination Classroom Server"
authors = ["The Reaper 3.0"]
license = "GNU GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.0.1"
discord = "^2.0.0"
python-dotenv = "^0.21.0"
pymongo = "^4.3.2"

[tool.poetry.group.dev.dependencies]
"discord.py" = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
